relenium {relenium}	R Documentation

Relenium: Selenium for R

Description

Selenium is a popular suite of tools specifically for automating web browsers. Selenium allows users to write tests in a number of popular programming languages including Java. The relenium package is intended to provide access to Selenium and use it to boost web navigation and to ease the path for web scraping in R.

Relenium classes

The relenium package is implemented with Reference Classes (see ReferenceClasses for more details). No knowledge of them is required, but the use of relenium may seem a little bit different from standard R. There are two classes: firefoxClass and webElementClass. The first one, represents a Firefox web browser and the second one an HTML element. There are methods in each class to navigate, select, find elements and to extract data.

Bugs and new features

To report technical problems with relenium, please file an issue on Github with a short reproducible example. Feedback is welcomed and any new features should be filed as issues as well.

[Package relenium version 0.3.0 Index]